Change to the algorithm and small GUI changes
In order to get elected, the criteria for a candidate to haves fewer than half the total ownvotes has changed to fewer than a third so people more hated have less of a chance of getting elected. The UI now has headers for number and candidate for the table and the import option will automatically refresh the page.